+++ NEWS.md
@@ -12,6 +12,7 @@ version 1.0.8
 * fix bug in `LANDISDisp()`: skip dispersal when `src` or `rcv` data.tables are empty.
 * initial cohort biomass (calculated in `.initiateNewCohorts`) can now be a fixed integer or, as before, calculated using the LANDIS-II Biomass Succession Extension v3.2.1 approach (Scheller & Miranda 2015).
 * fix bug in `prepInputsFireYear`.
+* drop single-level factor terms in `statsModel`.
 
 version 1.0.7
 =============